feat: ProgressDisplay 클래스 통합 및 UI 테스트 개선 - ProgressController 클래스를 ProgressDisplay로 통합하여 진행 상황 표시 기능을 개선 - 수동 새로 고침 옵션 추가로 유연한 진행 상황 제어 가능 - 기존의 UpdatableProgressReview 클래스 제거 및 새로운 enhanced_progress_review 메서드로 대체 - ReviewDisplay UI 테스트 스크립트 업데이트 및 새로운 테스트 케이스 추가
